child entity delete bug fixed

+++ outstack-spring/src/main/java/com/cloudimpl/outstack/spring/repo/PostgresRepositoryFactory.java
@@ -226,11 +226,13 @@ public class PostgresRepositoryFactory implements EventRepositoryFactory {
         }
     }
 
-    protected int deleteEventsByTrn(Connection conn, String tableName, String tenantId, String trn) {
+    protected int deleteEventsByTrn(Connection conn, String tableName, String tenantId, String trn, String eventOwner, String eventOwnerId) {
         createTenantIfNotExist(tableName, tenantId);
-        try ( PreparedStatement stmt = conn.prepareStatement("delete from " + tableName + " where tenantId = ? and trn = ?")) {
+        try ( PreparedStatement stmt = conn.prepareStatement("delete from " + tableName + " where tenantId = ? and trn = ? and eventOwner = ? and eventOwnerId = ?")) {
             stmt.setString(1, tenantId);
             stmt.setString(2, trn);
+            stmt.setString(3, eventOwner);
+            stmt.setString(4, eventOwnerId);
             return stmt.executeUpdate();
         } catch (SQLException ex) {
             throw new PostgresException(ex);
